[iPhone kettle] Add initial iPhone app
The iPhone app toggles a display based on
if the grid is good or not. It currently does
not interact with the webserver.

The app checks the users location, does a reverse
lookup to find the country the user is in, and warns
the user if he or she is not in the UK.