[iOS] compressed photos and load pictures
Compressed the PNGs differently to save space.
Now they are compressed with PNG-8 instead of
PNG-24. Seems to be a big win, but a loss in
reproduced colours. This doesn't seem to be
a problem as the palette I use is quite limited.
But you can see it a bit in the load screen images
where I used a gradient. The gradient becomes
a bit shitty, because there aren't enough colours
to represent all the different shades correctly.

I now also use smaller images for animations.
This is another small win in terms of space saving.