[iOS] Make a pluggable grid monitor infrastructure
Now, if a tactic fails, the app will try a different one.
This is kind of prematurely optimized in the way it is written,
but at least it looks relatively clean.
Have no internet on the plane, so cannot actually test,
but will look later.

The can I turn it on tactic has not been implemented yet.