[iOS] Bugfix plus indifferent blinds
Bugfix:
When loading the application, the correct orientation would not be
determined reliably. This caused the images for portrait and landscape
view to overlap and cause strange graphical artifacts.

Indifferent blinds:
Now an orangy yellowish blind is used until the grid state has
been determined. This is more appropriate than showing a green
blind!