[ios] bugfix
Fixed a bug where the kettle app would say the grid was
all good if it couldn't connect to the server.

I will not submit a new binary as that will put us at the back
of the update queue. Instead I will wait until we have been
accepted, and then submit an update.