[iOS] use crashlytics instead of crittercism
The reason for the choice is that crashalytics
uploads the dSym file automatically on each build,
which means that we get symbolicated reports all the
time automatically! A million times better!
And they are a much slicker startup anyway...
Wonder what their price plan will be like once
they are out of beta...